realocando arquivo de controller em sua pasta e renomenado funcoes de acesso

+++ admin/usuario/index.php
@@ -5,26 +5,29 @@ include_once"../../includes/util.php";
 $util = new  Util();
 $util->conectar();
 
-function listar(){
-	
-global $util;
-$util->lista_paginacao(MYSQL_BASE_USUARIOS, $_GET['pg']);
+function listar(){	
+	global $util;
+	$util->lista_paginacao(MYSQL_BASE_USUARIOS, $_GET['pg']);
 
-$sql = "SELECT * FROM ".MYSQL_BASE_USUARIOS." ORDER BY id DESC LIMIT ".$util->get_ini_reg_paginacao().",". $util->get_qtd_reg_paginacao();
-$res = mysql_query($sql);
-while($row = mysql_fetch_array($res)){
-		echo '<tr>';
-		echo '<td>&nbsp;</td>';
-		echo '<td><h3>'.$row['nome'].'</h3></td>';
-		echo '<td><h3>'.$row['apelido'].'</h3></td>';
-		echo '<td><h3>'.$row['email'].'</h3></td>';
-		echo '<td><h3>'.$row['senha'].'</h3></td>';
-		echo '<td><h3>'.$row['status'].'</h3></td>';
-		echo '<td><h3>'.$row['created_at'].'</h3></td>';
-		echo '<td><h3>'.$row['updated_at'].'</h3></td>';
-		echo '<td><a href="edit.php?id='.$row['id'].'" class=ico edit>Editar</a>&nbsp;&nbsp;<a href="delete.php?id='.$row['id'].'" class=ico del>Deletar</a></td>';
-		echo '</tr>';
-	}		
+	$sql = "SELECT * FROM ".MYSQL_BASE_USUARIOS." ORDER BY id DESC LIMIT ".$util->get_ini_reg_paginacao().",". 	$util->get_qtd_reg_paginacao();
+	$res = mysql_query($sql);
+	if(mysql_num_rows($res) > 0){
+		while($row = mysql_fetch_array($res)){
+			echo '<tr>';
+			echo '<td>&nbsp;</td>';
+			echo '<td><h3>'.$row['nome'].'</h3></td>';
+			echo '<td><h3>'.$row['apelido'].'</h3></td>';
+			echo '<td><h3>'.$row['email'].'</h3></td>';
+			echo '<td><h3>'.$row['senha'].'</h3></td>';
+			echo '<td><h3>'.$row['status'].'</h3></td>';
+			echo '<td><h3>'.$row['created_at'].'</h3></td>';
+			echo '<td><h3>'.$row['updated_at'].'</h3></td>';
+			echo '<td><a href="edit.php?id='.$row['id'].'" class=ico edit>Editar</a>&nbsp;&nbsp;<a href="delete.php?id='.$row['id'].'" 	class=ico del>Deletar</a></td>';
+			echo '</tr>';
+		}	
+	} else{
+		return false;
+	}	
 }
 ?>
 <div>
